+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## master
 
-- FIX: Put all errors into the AccessTokenAgent namespace 
+- Add `http_auth_header` method to the connector, since this is the most
+  common use case
+- Deprecate the `authenticate` method in favor of the new `token` method
+- Put all errors into the AccessTokenAgent namespace
 
 ## 3.1.1